Restructure repo layout
