+++ index.html
@@ -181,7 +181,7 @@
 <body>
     <div class="container">
         <div class="header">
-            <img src="skywater_logo.png" alt="Skywater Logo">
+            <img src="assets/skywater_logo.png" alt="Skywater Logo">
             <h1>VibeCheck Pro</h1>
             <p>Professional Vibration Analysis Tool</p>
         </div>
