+++ vibecheck/vc_generate_pdf.py
